# internal/routes/routes.go
package routes

import (
	"net/http"

	"github.com/paiva/SkillBridge/Backend/config"
	"github.com/paiva/SkillBridge/Backend/internal/handlers"
	"github.com/paiva/SkillBridge/Backend/internal/middleware"

	"github.com/gin-gonic/gin"

	ginSwagger "github.com/swaggo/gin-swagger"
    swaggerFiles "github.com/swaggo/files"
)

func Setup(r *gin.Engine) {

	// Swagger only in development/testing — acesso em /swagger/index.html
	if config.AppConfig.Env != "production" {
		r.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))
	}

	// Admin dashboard — first visit: /admin-dashboard?key=<ADMIN_SECRET_KEY>
	// After that the browser uses a session cookie automatically.
	r.GET("/admin-dashboard", func(c *gin.Context) {
		secret := config.AppConfig.AdminSecretKey
		if secret == "" {
			c.AbortWithStatus(http.StatusServiceUnavailable)
			return
		}

		// Check cookie first
		cookie, err := c.Cookie("admin_session")
		validCookie := err == nil && cookie == secret

		// Check query param
		queryKey := c.Query("key")
		validKey := queryKey != "" && queryKey == secret

		if !validCookie && !validKey {
			c.AbortWithStatus(http.StatusForbidden)
			return
		}

		// Set/renew cookie (7 days, HttpOnly, SameSite=Strict)
		c.SetCookie("admin_session", secret, 7*24*3600, "/", "", c.Request.TLS != nil, true)

		handlers.AdminDashboard(c)
	})


	// Servir imagens guardadas localmente
	r.Static("/uploads", config.AppConfig.UploadsDir)

	// CORS - permite pedidos do frontend Angular
	r.Use(func(c *gin.Context) {
		origin := c.Request.Header.Get("Origin")
		allowedOrigin := config.AppConfig.FrontendURL
		if origin == allowedOrigin || config.AppConfig.Env == "development" {
			if config.AppConfig.Env == "development" {
				c.Header("Access-Control-Allow-Origin", "*")
			} else {
				c.Header("Access-Control-Allow-Origin", allowedOrigin)
			}
		}
		c.Header("Access-Control-Allow-Methods", "GET, POST, PUT, DELETE, OPTIONS")
		c.Header("Access-Control-Allow-Headers", "Content-Type, Authorization, X-Admin-Key")
		c.Header("Access-Control-Allow-Credentials", "true")
		// Required for Firebase signInWithPopup to work across same origin
		c.Header("Cross-Origin-Opener-Policy", "same-origin-allow-popups")
		c.Header("Cross-Origin-Embedder-Policy", "unsafe-none")
		if c.Request.Method == "OPTIONS" {
			c.AbortWithStatus(204)
			return
		}
		c.Next()
	})

	api := r.Group("/api")

	// --------------------------------------------------
	// ROTAS PROTEGIDAS (requerem token Firebase)
	// --------------------------------------------------
	protected := api.Group("")
	protected.Use(middleware.AuthRequired())
	{
		// Utilizadores — rotas específicas ANTES do wildcard :id
		protected.POST("/users/register", handlers.RegisterUser)
		protected.GET("/users/me", handlers.GetMyProfile)
		protected.PUT("/users/me", handlers.UpdateProfile)
		protected.DELETE("/users/me", handlers.DeleteMyProfile)
		// Claim an anonymous guest session into the authenticated profile
		protected.POST("/users/me/claim-guest-session", handlers.ClaimGuestSession)

		// Seguidores
		protected.POST("/users/:id/follow", handlers.FollowUser)
		protected.DELETE("/users/:id/follow", handlers.UnfollowUser)
		protected.GET("/users/:id/follow/status", handlers.GetFollowStatus)

		// Skills do utilizador
		protected.POST("/users/me/skills", handlers.AddUserSkill)
		protected.DELETE("/users/me/skills", handlers.RemoveUserSkill)

		// Projetos
		protected.POST("/projects", handlers.CreateProject)
		protected.PUT("/projects/:id", handlers.UpdateProject)
		protected.DELETE("/projects/:id", handlers.DeleteProject)
		protected.PUT("/projects/:id/status", handlers.UpdateProjectStatus)
		protected.POST("/projects/:id/join", handlers.JoinProject)
		protected.GET("/projects/:id/applications", handlers.GetApplications)
		protected.PUT("/projects/:id/applications/:member_id", handlers.RespondApplication)
		protected.DELETE("/projects/:id/members/:member_id", handlers.RemoveProjectMember)
		protected.POST("/projects/:id/owners", handlers.AddProjectOwner)
		protected.POST("/projects/:id/roles", handlers.CreateProjectRole)
		protected.DELETE("/projects/:id/roles/:role_id", handlers.DeleteProjectRole)

		// Avaliações
		protected.POST("/reviews", handlers.CreateReview)

		// Upload de imagens
		protected.POST("/upload/image", handlers.UploadImage)
		protected.DELETE("/upload/image", handlers.DeleteImage)

		// Mensagens privadas E2E cifradas
		protected.PUT("/messages/keys", handlers.RegisterPublicKey)
		protected.GET("/messages/keys/:user_id", handlers.GetPublicKey)
		protected.GET("/messages/unread-count", handlers.GetUnreadCount)
		protected.POST("/conversations", handlers.StartOrGetConversation)
		protected.GET("/conversations", handlers.ListConversations)
		protected.POST("/conversations/:id/messages", handlers.SendMessage)
		protected.GET("/conversations/:id/messages", handlers.GetMessages)
		protected.PUT("/conversations/:id/read", handlers.MarkRead)
	}

	// --------------------------------------------------
	// ROTAS PÚBLICAS (sem autenticação)
	// — registadas DEPOIS das rotas /users/me protegidas
	//   para que o wildcard :id não as intercete primeiro
	// --------------------------------------------------
	public := api.Group("")
	{
		// Utilizadores (público) — wildcard :id deve vir DEPOIS de /me
		public.GET("/users/:id", handlers.GetUserByID)
		public.GET("/users/:id/reviews", handlers.GetUserReviews)
		public.GET("/users/:id/followers", handlers.GetFollowers)
		public.GET("/users/:id/following", handlers.GetFollowing)
		public.GET("/users/:id/follow/counts", handlers.GetFollowCounts)
		// Mot-de-passe (público — não requer token)
		public.POST("/users/password-reset", handlers.RequestPasswordReset)

		// Projetos
		public.GET("/projects", handlers.GetProjects)
		public.GET("/projects/:id", handlers.GetProjectByID)
		public.GET("/projects/:id/members", handlers.GetProjectMembers)

		// Skills (leitura pública)
		public.GET("/skills", handlers.ListSkills)

		// Universidades/cursos estáticos
		public.GET("/universities", handlers.ListUniversities)
		public.GET("/universities/courses", handlers.ListCoursesByUniversity)

		// Guest onboarding sessions (anónimo)
		public.POST("/guest/session", handlers.CreateGuestSession)
		public.GET("/guest/session/:token", handlers.GetGuestSession)
		public.GET("/guest/stats", handlers.GetPlatformStats)

		// Donativos
		public.POST("/donations/embedded-checkout", handlers.CreateEmbeddedCheckoutSession)
		public.POST("/donations/webhook", handlers.StripeWebhook)
		public.GET("/donations/stats", handlers.GetDonationStats)
	}

	// --------------------------------------------------
	// ROTAS DE ADMINISTRAÇÃO (requerem token + UID admin)
	// --------------------------------------------------
	admin := api.Group("/admin")
	admin.Use(middleware.AuthRequired(), middleware.AdminRequired())
	{
		admin.GET("/users", handlers.AdminListUsers)
		admin.GET("/users/:id", handlers.AdminGetUser)
		admin.DELETE("/users/:id", handlers.AdminDeleteUser)
		admin.GET("/projects", handlers.AdminListProjects)
		admin.DELETE("/projects/:id", handlers.AdminDeleteProject)
		admin.GET("/donations", handlers.AdminListDonations)
		admin.GET("/reviews", handlers.AdminListReviews)
		admin.PUT("/reviews/:id", handlers.AdminDecideReview)
		admin.DELETE("/reviews/:id", handlers.AdminDeleteReview)
	}
}